requirements.txt
langgraph>=0.2.0
langchain-openai>=0.3.0
langchain-core>=0.3.0
python-dotenv>=1.0.0
streamlit>=1.40.0
pandas>=2.0.0
pytest>=8.0.0
